requirements.txt
accelerate>=0.16.0,<1
transformers[torch]>=4.28.1,<5
torch>=1.13.1,<2
scipy
numpy
fastapi
geograpy3
nltk